requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.4
django-cors-headers==3.6.0
djangorestframework==3.12.2
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2020.5
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
